requirements.txt
flask
Flask-SQLAlchemy
flask-login
requests
flask_admin
boto3